MOD: MainController.java move to new package mainController
