+++ src/courseView/CourseSearchMenuView.java
@@ -3,7 +3,7 @@ package courseView;
 import java.util.ArrayList;
 
 import controllers.CourseController;
-import controllers.MainController;
+import mainController.MainController;
 import models.Course;
 import models.Teacher;
 import models.User;
