+++ src/registrationView/UserRegistrationView.java
@@ -1,6 +1,6 @@
 package registrationView;
 
-import controllers.MainController;
+import mainController.MainController;
 import tools.Alerts;
 import tools.Cnsl;
